refactor: rename src/auth.rs -> src/api_auth.rs
The module holds only ApiAuth (proxy -> Source API auth) and sits beside
backend_auth.rs (proxy -> storage backend). The generic name was the odd one
out; api_auth makes it parallel and self-describing. Module-path only — every
other reference goes through the crate-root `crate::ApiAuth` re-export.
